fix(home): rename component file name
